icu_locale::extensions::unicode

Struct Value

pub struct Value(/* private fields */);

A value used in a list of Keywords.

The value has to be a sequence of one or more alphanumerical strings separated by -. Each part of the sequence has to be no shorter than three characters and no longer than 8.

Examples

use icu::locale::extensions::unicode::{value, Value};
use writeable::assert_writeable_eq;

assert_writeable_eq!(value!("gregory"), "gregory");
assert_writeable_eq!(
    "islamic-civil".parse::<Value>().unwrap(),
    "islamic-civil"
);

// The value "true" has the special, empty string representation
assert_eq!(value!("true").to_string(), "");

Implementations

impl Value

pub fn try_from_str(s: &str) -> Result<Value, ParseError>

A constructor which str slice, parses it and produces a well-formed Value.

Examples

use icu::locale::extensions::unicode::Value;

Value::try_from_str("buddhist").expect("Parsing failed.");

pub fn try_from_utf8(code_units: &[u8]) -> Result<Value, ParseError>

See Self::try_from_str

pub const fn as_single_subtag(&self) -> Option<&Subtag>

Returns a reference to a single Subtag if the Value contains exactly one subtag, or None otherwise.

Examples

use core::str::FromStr;
use icu::locale::extensions::unicode::Value;

let value1 = Value::from_str("foo").expect("failed to parse a Value");
let value2 = Value::from_str("foo-bar").expect("failed to parse a Value");

assert!(value1.as_single_subtag().is_some());
assert!(value2.as_single_subtag().is_none());

pub fn into_single_subtag(self) -> Option<Subtag>

Destructs into a single Subtag if the Value contains exactly one subtag, or returns None otherwise.

Examples

use core::str::FromStr;
use icu::locale::extensions::unicode::Value;

let value1 = Value::from_str("foo").expect("failed to parse a Value");
let value2 = Value::from_str("foo-bar").expect("failed to parse a Value");

assert!(value1.into_single_subtag().is_some());
assert!(value2.into_single_subtag().is_none());

pub fn push_subtag(&mut self, subtag: Subtag)

Appends a subtag to the back of a Value.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};

let mut v = Value::default();
v.push_subtag(subtag!("foo"));
v.push_subtag(subtag!("bar"));
assert_eq!(v, "foo-bar");

pub fn subtag_count(&self) -> usize

Returns the number of subtags in the Value.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};

let mut v = Value::default();
assert_eq!(v.subtag_count(), 0);
v.push_subtag(subtag!("foo"));
assert_eq!(v.subtag_count(), 1);

pub const fn new_empty() -> Value

Creates an empty Value, which corresponds to a โtrueโ value.

Examples

use icu::locale::extensions::unicode::{value, Value};

assert_eq!(value!("true"), Value::new_empty());

pub fn is_empty(&self) -> bool

Returns true if the Value has no subtags.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};

let mut v = Value::default();
assert_eq!(v.is_empty(), true);

pub fn remove_subtag(&mut self, idx: usize) -> Option<Subtag>

Removes and returns the subtag at position index within the value, shifting all subtags after it to the left.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};
let mut v = Value::default();
v.push_subtag(subtag!("foo"));
v.push_subtag(subtag!("bar"));
v.push_subtag(subtag!("baz"));

assert_eq!(v.remove_subtag(1), Some(subtag!("bar")));
assert_eq!(v, "foo-baz");

pub fn get_subtag(&self, idx: usize) -> Option<&Subtag>

Returns a reference to a subtag at index.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};
let mut v = Value::default();
v.push_subtag(subtag!("foo"));
v.push_subtag(subtag!("bar"));
v.push_subtag(subtag!("baz"));

assert_eq!(v.get_subtag(1), Some(&subtag!("bar")));
assert_eq!(v.get_subtag(3), None);

pub fn from_vec_unchecked(input: Vec<Subtag>) -> Value

A constructor which takes a pre-sorted list of Value elements.

Examples

use icu::locale::extensions::unicode::Value;
use icu::locale::subtags::subtag;

let subtag1 = subtag!("foobar");
let subtag2 = subtag!("testing");
let mut v = vec![subtag1, subtag2];
v.sort();
v.dedup();

let value = Value::from_vec_unchecked(v);

Notice: For performance- and memory-constrained environments, it is recommended for the caller to use binary_search instead of sort and dedup.

impl Value

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient [Writeable] implementation. However, in order to avoid allocating a string, it is more efficient to use [Writeable] directly.